+++ CassiaWindowList@klangman/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 2.4.3
+
+* Register a hotkey for "0" when using smart numeric hotkeys which is used as the 10th hotkey
+* Add a "Button position priority list" option that allows uses set a persistent window list  button order using drag-and-drop and a context menu option called "Prioritize window list position"
+* Remove the thumbnail menu when opening a context menu for a window list button
+* Added a "Mouse" tab to the config window, this is done to make room for new options in the general tab
+
 ## 2.4.2
 
 * Renamed the "Number label" to "Icon overlay label" since it no longer only shows numbers